--- a/clients/src/main/java/org/apache/kafka/common/network/ByteBufferSend.java
+++ b/clients/src/main/java/org/apache/kafka/common/network/ByteBufferSend.java
@@ -27,9 +27,9 @@ import java.nio.channels.GatheringByteChannel;
 public class ByteBufferSend implements Send {
 
     private final String destination;
-    private final int size;
+    private final long size;
     protected final ByteBuffer[] buffers;
-    private int remaining;
+    private long remaining;
     private boolean pending = false;
 
     public ByteBufferSend(String destination, ByteBuffer... buffers) {
@@ -38,6 +38,13 @@ public class ByteBufferSend implements Send {
         for (ByteBuffer buffer : buffers)
             remaining += buffer.remaining();
         this.size = remaining;
+    }
+
+    public ByteBufferSend(String destination, ByteBuffer[] buffers, long size) {
+        this.destination = destination;
+        this.buffers = buffers;
+        this.size = size;
+        this.remaining = size;
     }
 
     @Override