--- a/clients/src/main/java/org/apache/kafka/common/protocol/SendBuilder.java
+++ b/clients/src/main/java/org/apache/kafka/common/protocol/SendBuilder.java
@@ -41,29 +41,19 @@ import java.util.Queue;
  * for example usage.
  */
 public class SendBuilder implements Writable {
-    private final Queue<Send> sends = new ArrayDeque<>(1);
-    private final List<ByteBuffer> buffers = new ArrayList<>();
     private final ByteBuffer buffer;
     private final String destinationId;
+
+    private final Queue<Send> sends = new ArrayDeque<>(1);
+    private long sizeOfSends = 0;
+
+    private final List<ByteBuffer> buffers = new ArrayList<>();
+    private long sizeOfBuffers = 0;
 
     SendBuilder(String destinationId, int size) {
         this.destinationId = destinationId;
         this.buffer = ByteBuffer.allocate(size);
         this.buffer.mark();
-    }
-
-    private void flushPendingBuffer() {
-        int latestPosition = buffer.position();
-        buffer.reset();
-
-        if (latestPosition > buffer.position()) {
-            buffer.limit(latestPosition);
-            buffers.add(buffer.slice());
-
-            buffer.position(latestPosition);
-            buffer.limit(buffer.capacity());
-            buffer.mark();
-        }
     }
 
     @Override
@@ -110,7 +100,7 @@ public class SendBuilder implements Writable {
     @Override
     public void writeByteBuffer(ByteBuffer buf) {
         flushPendingBuffer();
-        buffers.add(buf.duplicate());
+        addBuffer(buf.duplicate());
     }
 
     @Override
@@ -123,12 +113,19 @@ public class SendBuilder implements Writable {
         ByteUtils.writeVarlong(i, buffer);
     }
 
-    private void flushPendingSend() {
-        if (!buffers.isEmpty()) {
-            ByteBuffer[] byteBufferArray = buffers.toArray(new ByteBuffer[0]);
-            sends.add(new ByteBufferSend(destinationId, byteBufferArray));
-            buffers.clear();
-        }
+    private void addBuffer(ByteBuffer buffer) {
+        buffers.add(buffer);
+        sizeOfBuffers += buffer.remaining();
+    }
+
+    private void addSend(Send send) {
+        sends.add(send);
+        sizeOfSends += send.size();
+    }
+
+    private void clearBuffers() {
+        buffers.clear();
+        sizeOfBuffers = 0;
     }
 
     /**
@@ -139,24 +136,45 @@ public class SendBuilder implements Writable {
      */
     @Override
     public void writeRecords(BaseRecords records) {
-        flushPendingBuffer();
-
         if (records instanceof MemoryRecords) {
-            buffers.add(((MemoryRecords) records).buffer());
+            flushPendingBuffer();
+            addBuffer(((MemoryRecords) records).buffer());
         } else {
             flushPendingSend();
-            sends.add(records.toSend(destinationId));
+            addSend(records.toSend(destinationId));
+        }
+    }
+
+    private void flushPendingSend() {
+        flushPendingBuffer();
+        if (!buffers.isEmpty()) {
+            ByteBuffer[] byteBufferArray = buffers.toArray(new ByteBuffer[0]);
+            addSend(new ByteBufferSend(destinationId, byteBufferArray, sizeOfBuffers));
+            clearBuffers();
+        }
+    }
+
+    private void flushPendingBuffer() {
+        int latestPosition = buffer.position();
+        buffer.reset();
+
+        if (latestPosition > buffer.position()) {
+            buffer.limit(latestPosition);
+            addBuffer(buffer.slice());
+
+            buffer.position(latestPosition);
+            buffer.limit(buffer.capacity());
+            buffer.mark();
         }
     }
 
     public Send build() {
-        flushPendingBuffer();
         flushPendingSend();
 
         if (sends.size() == 1) {
             return sends.poll();
         } else {
-            return new MultiRecordsSend(destinationId, sends);
+            return new MultiRecordsSend(destinationId, sends, sizeOfSends);
         }
     }