--- a/clients/src/main/java/org/apache/kafka/common/record/BaseRecords.java
+++ b/clients/src/main/java/org/apache/kafka/common/record/BaseRecords.java
@@ -30,5 +30,5 @@ public interface BaseRecords {
      * Encapsulate this {@link BaseRecords} object into {@link RecordsSend}
      * @return Initialized {@link RecordsSend} object
      */
-    RecordsSend toSend(String destination);
+    RecordsSend<? extends BaseRecords> toSend(String destination);
 }