@hachikuji 2020-11-19
@chia7712 To be honest, I'm not sure. I think it makes sense in general to try and batch the writes together to reduce number of calls to the kernel in any case, but I am not sure why it has such a big impact locally. I also didn't see anything like this when I was testing in on linux, but I didn't feel great about the difference anyway.